+++ test/browsercontext.spec.js
@@ -15,10 +15,9 @@
  */
 
 const utils = require('./utils');
-const puppeteer = utils.requireRoot('index');
 const {TimeoutError} = utils.requireRoot('Errors');
 
-module.exports.addTests = function({testRunner, expect}) {
+module.exports.addTests = function({testRunner, expect, puppeteer}) {
   const {describe, xdescribe, fdescribe} = testRunner;
   const {it, fit, xit} = testRunner;
   const {beforeAll, beforeEach, afterAll, afterEach} = testRunner;
