+++ test/golden-chromium/grid-cell-0.png
