+++ test/golden-chromium/grid-cell-1.png
