+++ test/golden-chromium/grid-cell-2.png
