+++ test/golden-chromium/grid-cell-3.png
