+++ test/golden-chromium/jscoverage-involved.txt
