+++ test/golden-chromium/mock-binary-response.png
