+++ test/golden-chromium/nested-frames.txt
