+++ test/golden-chromium/screenshot-clip-odd-size.png
