+++ test/golden-chromium/screenshot-clip-rect.png
