+++ test/golden-chromium/screenshot-element-bounding-box.png
