+++ test/golden-chromium/screenshot-element-fractional-offset.png
