+++ test/golden-chromium/screenshot-element-fractional.png
