+++ test/golden-chromium/screenshot-element-larger-than-viewport.png
