+++ test/golden-chromium/screenshot-element-padding-border.png
