+++ test/golden-chromium/screenshot-element-rotate.png
