+++ test/golden-chromium/screenshot-element-scrolled-into-view.png
