+++ test/golden-chromium/screenshot-offscreen-clip.png
