+++ test/golden-chromium/screenshot-sanity.png
