+++ test/golden-chromium/transparent.png
