+++ test/golden-chromium/white.jpg
