+++ test/ignorehttpserrors.spec.js
@@ -14,10 +14,7 @@
  * limitations under the License.
  */
 
-const utils = require('./utils');
-const puppeteer = utils.requireRoot('index.js');
-
-module.exports.addTests = function({testRunner, expect, defaultBrowserOptions}) {
+module.exports.addTests = function({testRunner, expect, defaultBrowserOptions, puppeteer}) {
   const {describe, xdescribe, fdescribe} = testRunner;
   const {it, fit, xit} = testRunner;
   const {beforeAll, beforeEach, afterAll, afterEach} = testRunner;
